SystemCode/backend/requirements.txt
numpy~=1.18.1
matplotlib~=3.1.3
imageio~=2.6.1
torchvision~=0.7.0
django~=2.2.6
djangorestframework~=3.11.1
pandas~=1.0.1
opencv-python~=4.4.0.44
scikit-learn~=0.22.1
tensorflow~=2.2.0
scipy~=1.4.1
numba~=0.48.0
pillow~=7.0.0
scikit-image~=0.16.2